app/requirements.txt
Flask==3.1.2
mlflow-skinny==2.22.0
cloudpickle==3.1.0
pandas==2.3.3
scikit-learn==1.7.2
xgboost==2.1.4
gunicorn
